+++ src/core/pre.js
@@ -9,14 +9,6 @@ function sleep(ms) {
   return new Promise((res) => setTimeout(res, ms));
 }
 
-function decodeHexString(s) {
-  const result = new Uint8Array(s.length / 2);
-  for (let i = 0; i < s.length / 2; i++) {
-    result[i] = Number.parseInt(s.slice(2 * i, 2 * i + 2), 16);
-  }
-  return result;
-}
-
 function getTypeTag(x) {
   try {
     return Object.prototype.toString.call(x);
