+++ src/js/package.json
@@ -78,7 +78,7 @@
   },
   "scripts": {
     "test": "npm-run-all test:*",
-    "test:unit": "cross-env TEST_NODE=1 ts-mocha -p tsconfig.test.json test/unit/**/*.test.*",
+    "test:unit": "cross-env TEST_NODE=1 ts-mocha --node-option=experimental-loader=./test/loader.mjs -p tsconfig.test.json test/unit/**/*.test.*",
     "test:node": "cross-env TEST_NODE=1 mocha test/integration/**/*.test.js",
     "test:browser": "mocha test/integration/**/*.test.js",
     "tsc": "tsc --noEmit",
