+++ src/js/test/unit/wat/invoke_dd.wat
