+++ src/js/test/unit/wat/invoke_vd.wat
